fix: move useSearch hook to frontend package
