+++ packages/react/src/hooks/index.ts
@@ -22,5 +22,4 @@ export * from "./useTemplateModule";
 export * from "./useTheme";
 export * from "./useTokenLists";
 export * from "./useWatchData";
-export * from "./useSearch";
 export * from "./useResetPageScroll";
